Conflicts:
	cps_pp/include/util/dirac_op.h
	cps_pp/include/util/lattice.h
	cps_pp/src/alg/alg_tcharge/alg_tcharge.C
	cps_pp/src/alg/alg_threept/alg_threept.C
	cps_pp/src/util/lattice/f_bfm/f_bfm.C
	cps_pp/src/util/lattice/f_wilsonTm/f_wilsonTm.C
	cps_pp/src/util/lattice/fix_gauge/fix_gauge.C
	cps_pp/src/util/lattice/g_wilson/g_wilson.C
	cps_pp/src/util/lattice/lattice_base/lattice_base.C

+++ cps_pp/src/util/parallel_transport/pt_base/pt_base.C
@@ -37,7 +37,7 @@ CPS_START_NAMESPACE
 // Initialize static variables
 //------------------------------------------------------------------
 int ParTrans::scope_lock = 0;
-int ParTrans::PTflops = 0;
+double ParTrans::PTflops = 0;
 int ParTrans:: bc[4] = {0,0,0,0};	// boundary condition on this node
 int ParTrans::node_sites[5] = {0,0,0,0,0};
